pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "newsapi-python>=0.2.7",
    "nltk>=3.9.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "streamlit>=1.44.1",
    "trafilatura>=2.0.0",
    "yfinance>=0.2.55",
]
